refactor: book ثxisting validation

+++ Backend/app/Http/Controllers/BookController.php
@@ -20,16 +20,27 @@ class BookController extends Controller
     public function index()
     {
         $books = Book::all();
+        if($books->isEmpty()){
+            return response()->json(["message"=>"There is no books"],201);
+        }
         return response()->json(["books"=>$books],201);
     }
     public function show($slug)
     {
         $book=Book::where('slug',$slug)->get();
+        if(!$book)
+        {
+            return response()->json(['error' => 'Book not found']);
+        }
         return response()->json(["book"=>$book],201);
     }
     public function destroy($id)
     {
         $book = Book::find($id);
+        if(!$book)
+        {
+            return response()->json(['error' => 'Book not found']);
+        }
         $book->delete();
 
         return response()->json(['message' => 'Book deleted successfully']);
